requirements.txt
asttokens==2.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2023.7.22
charset-normalizer==2.0.12
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.4.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.3.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2021.1
pyzmq==25.1.1
regex==2021.4.4
requests==2.26.0
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
ticktick-py==2.0.3
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==1.26.7
wcwidth==0.2.8
websocket-client==1.6.4
wheel==0.41.2
zipp==3.17.0
